+++ web/src/routes/cabinet/files/index.tsx
@@ -11,7 +11,6 @@ function RootDirectoryContent() {
   const ensureDirectoryDecrypted = useDocumentsStore((s) => s.ensureDirectoryDecrypted);
   const viewMode = useDocumentsStore((s) => s.viewMode);
   const downloadFile = useDocumentsStore((s) => s.downloadFile);
-  const downloadingUris = useDocumentsStore((s) => s.downloadingUris);
   const items = useDocumentsStore(useShallow((s) => s.itemsForDirectory(null)));
 
   useEffect(() => {
@@ -31,7 +30,6 @@ function RootDirectoryContent() {
       viewMode={viewMode}
       onOpen={handleOpen}
       onDownload={(uri) => void downloadFile(uri)}
-      downloadingUris={downloadingUris}
     />
   );
 }